requirements.txt
streamlit==1.37.1
python-dotenv==1.0.1
feedparser==6.0.11
pydub==0.25.1
requests==2.32.3
openai==1.40.2
httpx==0.27.0
python-slugify==8.0.4
langcodes==3.4.0
pyyaml==6.0.2
watchdog==4.0.1
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1